Preventing tests from failing on API 23 (permissions)

+++ app/src/androidTest/java/me/angrybyte/contactsgenerator/test/app/OperationTest.java
@@ -1,13 +1,20 @@
 
 package me.angrybyte.contactsgenerator.test.app;
 
+import android.Manifest;
 import android.content.OperationApplicationException;
+import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
+import android.os.Build;
 import android.os.RemoteException;
+import android.support.v4.app.ActivityCompat;
 import android.test.ActivityInstrumentationTestCase2;
 import android.test.suitebuilder.annotation.MediumTest;
 import android.test.suitebuilder.annotation.SmallTest;
 import android.text.TextUtils;
+import android.util.Log;
+
+import java.util.List;
 
 import me.angrybyte.contactsgenerator.MainActivity;
 import me.angrybyte.contactsgenerator.R;
@@ -15,13 +22,12 @@ import me.angrybyte.contactsgenerator.api.ContactOperations;
 import me.angrybyte.contactsgenerator.api.Operations;
 import me.angrybyte.contactsgenerator.parser.data.Person;
 
-import java.util.List;
-
 /**
  * A test case used for testing app's core functionality. Note that all methods are returning {@code void} and asserting result values.
  */
 public class OperationTest extends ActivityInstrumentationTestCase2<MainActivity> {
 
+    public static final String TAG = OperationTest.class.getSimpleName();
     public static final String HTTP_TEST_URL = "http://pastebin.com/raw.php?i=VS10rPNJ";
     public static final String HTTP_TEST_CONTENT = "TESTING... OK!";
     public static final int DEFAULT_KEY_LENGTH = 19;
@@ -93,16 +99,30 @@ public class OperationTest extends ActivityInstrumentationTestCase2<MainActivity
 
     @MediumTest
     public void testFContactStorage() throws RemoteException, OperationApplicationException {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && !hasPermission(Manifest.permission.WRITE_CONTACTS)) {
+            Log.w(TAG, "Cannot run this test without the user permission");
+            return;
+        }
+
         ContactOperations contacts = new ContactOperations(mActivity);
         List<Person> persons = mApi.getPersons(5, Operations.BOTH);
+        int iterations = 0;
         for (Person person : persons) {
             person.setImage(mApi.fetchImage(person));
             contacts.storeContact(person);
+            iterations++;
         }
+
+        assertEquals("Failed to store all contacts. Stored " + iterations + " out of " + persons.size(), iterations, persons.size());
     }
 
     @MediumTest
     public void testGDeleteAllContacts() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && !hasPermission(Manifest.permission.WRITE_CONTACTS)) {
+            Log.w(TAG, "Cannot run this test without the user permission");
+            return;
+        }
+
         ContactOperations contacts = new ContactOperations(mActivity);
         boolean deleted = contacts.deleteContacts(null);
         assertTrue("Contacts not deleted", deleted);
@@ -110,9 +130,18 @@ public class OperationTest extends ActivityInstrumentationTestCase2<MainActivity
 
     @MediumTest
     public void testHDeleteExampleContacts() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && !hasPermission(Manifest.permission.WRITE_CONTACTS)) {
+            Log.w(TAG, "Cannot run this test without the user permission");
+            return;
+        }
+
         ContactOperations contacts = new ContactOperations(mActivity);
         boolean deleted = contacts.deleteContacts("example.com");
         assertTrue("Contacts not deleted", deleted);
+    }
+
+    private boolean hasPermission(String which) {
+        return ActivityCompat.checkSelfPermission(mActivity, which) == PackageManager.PERMISSION_GRANTED;
     }
 
     @Override
